+++ packages/core-config/src/SecureURL.ts
@@ -1,68 +1,54 @@
 import util, { type InspectOptions } from 'node:util';
-import { sha256 } from './sha256';
+import { SecureString } from './SecureString';
 import { BaseObject, type InspectFunction } from './types';
-
-type UrlObject = {
-  href: string;
-  password?: string;
-  searchParams?: Record<string, string>;
-};
 
 /**
  * SecureURL obfuscates the password from the URL.
  */
 export class SecureURL extends BaseObject {
   readonly #value: URL;
+  readonly #password: SecureString | null;
 
   public get secretValue(): URL {
     return this.#value;
   }
 
-  private constructor(value: URL) {
+  private constructor(value: URL, secret: string | undefined) {
     super();
     this.#value = value;
+    this.#password = SecureString.from(value.password || null, secret);
   }
 
-  public static from(value: undefined | null): null;
-  public static from(value: URL): SecureURL;
-  public static from(value: URL | null | undefined): SecureURL | null {
-    if (value == null) {
-      return null;
+  static factory(secret: string | undefined): (value: URL) => SecureURL {
+    return (value: URL) => SecureURL.from(value, secret);
+  }
+
+  public static from<T extends URL | null | undefined>(value: T, secret?: string): T extends URL ? SecureURL : T {
+    if (value === null) {
+      return null as T extends URL ? SecureURL : T;
     }
-    return new SecureURL(value);
+    if (value === undefined) {
+      return undefined as T extends URL ? SecureURL : T;
+    }
+    return new SecureURL(value, secret) as T extends URL ? SecureURL : T;
   }
 
   override toString(): string {
-    const newUrl = new URL(this.#value.href);
-    if (newUrl.password) {
-      newUrl.password = sha256(newUrl.password);
-    }
-    return newUrl.href;
+    return this.toURL().href;
   }
 
-  override toJSON(): object {
+  toURL(): URL {
     const newUrl = new URL(this.#value.href);
-
-    const searchParams = new URLSearchParams(newUrl.searchParams);
-    newUrl.search = '';
-
-    let password: string | undefined;
-    if (newUrl.password) {
-      password = sha256(newUrl.password);
-      newUrl.password = '';
+    if (this.#password !== null) {
+      newUrl.password = this.#password.toString();
     }
-
-    const result: UrlObject = {
-      href: newUrl.href,
-    };
-    if (password !== undefined) {
-      result.password = password;
-    }
-    if (searchParams.size > 0) {
-      result.searchParams = Object.fromEntries(searchParams);
-    }
-    return result;
+    return newUrl;
   }
+
+  override toJSON(): string {
+    return JSON.stringify(this.toURL());
+  }
+
   override [util.inspect.custom](depth: number, options: InspectOptions, inspect: InspectFunction): string {
     if (depth < 0) {
       return '[SecureURL]';
@@ -70,6 +56,6 @@ export class SecureURL extends BaseObject {
     const newOptions = Object.assign({}, options, {
       depth: options.depth == null ? null : options.depth - 1,
     });
-    return inspect(this.toJSON(), newOptions);
+    return inspect(this.toURL(), newOptions);
   }
 }