+++ packages/core-config/src/index.ts
@@ -1,6 +1,7 @@
 import { SecureConnectionString } from './SecureConnectionString';
 import { SecureString } from './SecureString';
 import { SecureURL } from './SecureURL';
+import { createFactory } from './createFactory';
 import { defaultSecureKeys } from './defaults';
 
-export { SecureConnectionString, SecureString, SecureURL, defaultSecureKeys };
+export { SecureConnectionString, SecureString, SecureURL, createFactory, defaultSecureKeys };